chore: move BitwiseBuiltinRunner to own file
